refactor(renderer): move the main view's wiring out of the engine
