+++ src/modules/ui/dom.ts
@@ -37,13 +37,13 @@ import { t } from "@core/i18n";
 import { disconnectResizeObserver } from "@modules/lyrics/injectLyrics";
 import type { ThumbnailElement } from "@modules/lyrics/requestSniffer/NextResponse";
 import { getArtworkMetadata } from "@modules/lyrics/requestSniffer/requestSniffer";
-import { animEngineState, reflow, resetAnimEngineState, toMs } from "@modules/ui/animationEngine";
-import { lyricsElementAdded } from "@modules/ui/mainLyricsView";
+import { animEngineState, lyricsElementAdded, resetAnimEngineState } from "@modules/ui/mainLyricsView";
 import { getResumeScrollElement } from "@modules/ui/resumeScrollButton";
 import { getRequest, setRequest } from "@modules/unison/lyricsRequestTracker";
 import { getTrustTier } from "@modules/unison/trustTier";
 import type { UnisonLyricsRequest } from "@modules/unison/types";
 import { requestLyrics } from "@modules/unison/unisonApi";
+import { reflow, toMs } from "@renderer/util";
 import { log } from "@utils";
 import { generatePetName } from "@/core/keyIdentity";
 import { byId, deleteVote, type UnisonData, vote } from "../lyrics/providers/unison";
